Retain manually-set y-limits when plotting under "hold on".
Inside plot(), we usually have the techs suggest some y-limits so that
the function graph gets displayed in full; however, it is not
appropriate to use these if the y-limits have been set manually and
we're plotting onto a figure that has been held.

We _were_ checking for this scenario, but we were setting the limits
from the wrong variable when it occurred:  we should set them using the
values saved in yLimCurrent at the beginning of the routine, not those
in yLim, which contains the "suggested" values (possibly adjusted a bit
along the way for various reasons).

Closes #1815.